Pet-40_Create specified methods
