+++ src/main/java/org/model/Pet.java
@@ -1,22 +1,14 @@
 package org.model;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import javafx.application.Application;
-import javafx.scene.Group;
-import javafx.scene.Scene;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.stage.Stage;
 
-import org.dizitart.no2.objects.Id;
+
+import java.awt.image.BufferedImage;
 
 public class Pet{
     private String type;
-    private Image image;
+    private BufferedImage image;
     private String info;
 
-    public Pet(String type, String info, Image image) {
+    public Pet(String type, String info, BufferedImage image) {
         this.type = type;
         this.info=info;
         this.image=image;
@@ -27,10 +19,10 @@ public class Pet{
 
     public String getType(){ return this.type; }
     public String getInfo(){ return this.info; }
-    public Image getImage(){ return this.image; }
+    public BufferedImage getImage(){ return this.image; }
     public void setType(String type){ this.type=type; }
     public void setInfo(String info){ this.info=info; }
-    public void setImage(Image image){ this.image=image; }
+    public void setImage(BufferedImage image){ this.image=image; }
 
     public boolean equals(Object o){
         if(o instanceof Pet){
